CA.Evaluation: lots of renames
