+++ Agents/Xamarin.Interactive/CodeAnalysis/Compilation.cs
@@ -10,6 +10,7 @@ using System.Collections.Generic;
 
 using Newtonsoft.Json;
 
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.CodeAnalysis.Resolving;
 
 namespace Xamarin.Interactive.CodeAnalysis
