+++ Agents/Xamarin.Interactive/Serialization/InteractiveJsonSerializerSettings.cs
@@ -13,6 +13,7 @@ using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
 
 using Xamarin.Interactive.CodeAnalysis;
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.Remote;
 using Xamarin.Interactive.Representations.Reflection;
