+++ Clients/Xamarin.Interactive.Client/Workbook/Models/HiddenCodeCellView.cs
@@ -9,6 +9,7 @@ using System;
 using System.Globalization;
 
 using Xamarin.Interactive.CodeAnalysis;
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.CodeAnalysis.Models;
 using Xamarin.Interactive.Editor;
 using Xamarin.Interactive.Workbook.Views;
