+++ Clients/Xamarin.Interactive.Client/Workbook/Views/ICodeCellView.cs
@@ -9,6 +9,7 @@ using System;
 using System.Globalization;
 
 using Xamarin.Interactive.CodeAnalysis;
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.CodeAnalysis.Models;
 
 namespace Xamarin.Interactive.Workbook.Views
