+++ CodeAnalysis/Xamarin.Interactive.CodeAnalysis.Roslyn/RoslynCompilationWorkspace.cs
@@ -16,6 +16,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using Xamarin.Interactive.CodeAnalysis;
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.CodeAnalysis.Models;
 using Xamarin.Interactive.CodeAnalysis.Resolving;
 using Xamarin.Interactive.CodeAnalysis.Roslyn;
