+++ CodeAnalysis/Xamarin.Interactive.CodeAnalysis/CodeAnalysis/Evaluating/IEvaluationService.cs
@@ -10,7 +10,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Xamarin.Interactive.CodeAnalysis
+namespace Xamarin.Interactive.CodeAnalysis.Evaluating
 {
     /// <summary>
     /// Do not implement! This is just a temporary measure to allow
