+++ CodeAnalysis/Xamarin.Interactive.CodeAnalysis/CodeAnalysis/IWorkspaceService.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
+using Xamarin.Interactive.CodeAnalysis.Evaluating;
 using Xamarin.Interactive.CodeAnalysis.Models;
 using Xamarin.Interactive.CodeAnalysis.Resolving;
